requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
pillow>=9.0.0
opencv-python>=4.5.0
ultralytics>=8.0.0
matplotlib>=3.5.0
numpy>=1.21.0
sqlite3  # Usually included with Python